[JNI Example] call native in Java
b/372559388
